+++ src/types.ts
@@ -512,13 +512,18 @@ export interface ScreeningPolicyConfiguration {
     outboundTransactionDelay?: number;
 }
 
-export enum TravelRuleAction {
+export interface ScreeningTenantConfiguration {
+    disableBypass: boolean;
+    disableUnfreeze: boolean;
+}
+
+export enum ScreeningAction {
     screen = "SCREEN",
     pass = "PASS",
     freeze = "FREEZE"
 }
 
-export interface TravelRulePolicyRule {
+export interface ScreeningPolicyRuleResponse {
     sourceType?: string;
     sourceSubType?: string;
     destType?: string;
@@ -532,7 +537,16 @@ export interface TravelRulePolicyRule {
     amountUSD?: number;
     networkProtocol?: string;
     operation?: string;
-    action: TravelRuleAction;
+    action: ScreeningAction;
+}
+
+export interface ScreeningProviderConfigurationResponse {
+    direction?: TransactionDirection;
+    status?: ScreeningTransactionStatus;
+    amountUSD?: number;
+    amount?: number;
+    asset?: string;
+    action: ScreeningVerdict;
 }
 
 export enum PolicyApprovalStatus {
@@ -545,7 +559,7 @@ export enum TransactionDirection {
     outbound = "OUTBOUND"
 }
 
-export enum FbTravelRuleTransactionStatus {
+export enum ScreeningTransactionStatus {
     completed = "COMPLETED",
     pending = "PENDING",
     rejected = "REJECTED",
@@ -554,7 +568,7 @@ export enum FbTravelRuleTransactionStatus {
     blockingTimeExpired = "BLOCKING_TIME_EXPIRED",
 }
 
-export enum TravelRuleVerdict {
+export enum ScreeningVerdict {
     accept = "ACCEPT",
     reject = "REJECT",
     alert = "ALERT",
@@ -563,23 +577,39 @@ export enum TravelRuleVerdict {
     cancel = "CANCEL"
 }
 
-export interface TravelRuleRulesConfiguration {
-    direction?: TransactionDirection;
-    status?: FbTravelRuleTransactionStatus;
-    amountUSD?: number;
-    amount?: number;
-    asset?: string;
-    action: TravelRuleVerdict;
+export interface ScreeningConfigurationsResponse {
+    bypassScreeningDuringServiceOutages: boolean;
+    inboundTransactionDelay: number;
+    outboundTransactionDelay: number;
 }
 
-export interface TravelRulePolicy {
+export interface ScreeningPolicyResponse {
     tenantId?: string;
-    policy: TravelRulePolicyRule[];
+    policy: ScreeningPolicyRuleResponse;
     policyStatus?: PolicyApprovalStatus;
     isDefault: boolean;
     createDate?: Date;
     lastUpdate: Date;
 }
+
+export interface ScreeningPolicyRuleResponse {
+    sourceType?: string;
+    sourceSubType?: string;
+    destType?: string;
+    destSubType?: string;
+    destAddress?: string;
+    sourceId?: string;
+    destId?: string;
+    asset?: string;
+    baseAsset?: string;
+    amount?: number;
+    amountUSD?: number;
+    networkProtocol?: string;
+    operation?: string;
+    action: ScreeningAction;
+}
+
+export type ScreeningType = "travel_rule" | "aml";
 
 export enum Web3ConnectionFeeLevel {
     HIGH = "HIGH",
@@ -851,6 +881,8 @@ export interface TransactionFilter {
     sourceType?: PeerType;
     destType?: PeerType;
     sourceId?: string;
+    sourceWalletId?: string;
+    destWalletId?: string;
     destId?: string;
     sort?: "ASC" | "DESC";
 }
@@ -997,6 +1029,7 @@ export enum PeerType {
     EXCHANGE_ACCOUNT = "EXCHANGE_ACCOUNT",
     INTERNAL_WALLET = "INTERNAL_WALLET",
     EXTERNAL_WALLET = "EXTERNAL_WALLET",
+    CONTRACT = "CONTRACT",
     UNKNOWN = "UNKNOWN",
     NETWORK_CONNECTION = "NETWORK_CONNECTION",
     FIAT_ACCOUNT = "FIAT_ACCOUNT",
@@ -1162,7 +1195,7 @@ export interface AddressResponse {
     tag: string;
     type?: number;
     customerRefId?: number;
-    addressFormat: number;
+    addressFormat?: string;
     legacyAddress?: string;
     enterpriseAddress?: string;
     bip44AddressIndex?: string;
@@ -1278,6 +1311,35 @@ export interface User {
     role: string;
 }
 
+
+export type TRole =
+  | "ADMIN"
+  | "SIGNER"
+  | "EDITOR"
+  | "APPROVER"
+  | "VIEWER"
+  | "NON_SIGNING_ADMIN"
+  | "AUDITOR"
+  | "NCW_ADMIN"
+  | "NCW_SIGNER";
+
+interface BaseUser {
+    id: string;
+    enabled: boolean;
+    role: TRole;
+    status: string;
+}
+export interface ConsoleUser extends BaseUser {
+    firstName: string;
+    lastName: string;
+    email: string;
+    userType: "CONSOLE";
+}
+
+export interface ApiUser extends BaseUser {
+    name: string;
+    userType: "API";
+}
 export interface UsersGroup {
     id: string;
     name: string;
@@ -1513,6 +1575,10 @@ export interface Audit {
 export interface AuditsResponse {
     data: Audit[];
     total: number;
+}
+
+export interface AuditLogsResponse extends AuditsResponse {
+    cursor: string | null;
 }
 
 export interface ISystemMessageInfo {
@@ -1858,6 +1924,7 @@ export interface SmartTransfersTicketTerm {
     id: string;
     asset: string;
     amount: string;
+    amountUsd?: string;
     fromNetworkId: string;
     fromNetworkIdName?: string;
     toNetworkId: string;
@@ -1940,6 +2007,19 @@ export namespace NCW {
     export interface WalletInfo {
         walletId: string;
         enabled: boolean;
+    }
+
+    export class LatestBackupKey {
+        deviceId: string;
+        publicKey: string;
+        keyId: string;
+        algorithm: string;
+    }
+
+    export class LatestBackupResponse {
+        passphraseId: string;
+        createdAt: number;
+        keys: Array<LatestBackupKey>;
     }
 
     export interface GetWalletsPayload {