+++ CHANGELOG.md
@@ -2,11 +2,75 @@
 
 All notable changes to this project will be documented in this file. Dates are displayed in UTC.
 
-#### [v5.11.0](https://github.com/fireblocks/fireblocks-sdk-js/compare/v5.9.0...v5.11.0)
+#### [v5.18.0](https://github.com/fireblocks/fireblocks-sdk-js/compare/v5.17.0...v5.18.0)
+
+> 18 January 2024
+
+- feat: Add request interceptor and httpsAgent options [`#245`](https://github.com/fireblocks/fireblocks-sdk-js/pull/245)
+
+#### [v5.17.0](https://github.com/fireblocks/fireblocks-sdk-js/compare/v5.16.0...v5.17.0)
+
+> 7 January 2024
+
+- Audit logs api [`#240`](https://github.com/fireblocks/fireblocks-sdk-js/pull/240)
+
+#### [v5.16.0](https://github.com/fireblocks/fireblocks-sdk-js/compare/v5.15.1...v5.16.0)
+
+> 31 December 2023
+
+- fix: add missing params [`#241`](https://github.com/fireblocks/fireblocks-sdk-js/pull/241)
+
+#### [v5.15.1](https://github.com/fireblocks/fireblocks-sdk-js/compare/v5.15.0...v5.15.1)
+
+> 26 December 2023
+
+- Rename management paths [`#238`](https://github.com/fireblocks/fireblocks-sdk-js/pull/238)
+
+#### [v5.15.0](https://github.com/fireblocks/fireblocks-sdk-js/compare/v5.14.0...v5.15.0)
+
+> 24 December 2023
+
+- Support tenant's OTA configuration functionality [`#237`](https://github.com/fireblocks/fireblocks-sdk-js/pull/237)
+- feat: Add PeerType [`#227`](https://github.com/fireblocks/fireblocks-sdk-js/pull/227)
+
+#### [v5.14.0](https://github.com/fireblocks/fireblocks-sdk-js/compare/v5.13.0...v5.14.0)
+
+> 12 December 2023
+
+- Fix empty querystring scenario [`#235`](https://github.com/fireblocks/fireblocks-sdk-js/pull/235)
+-  Exposing external aml endpoints for screening setting [`#226`](https://github.com/fireblocks/fireblocks-sdk-js/pull/226)
+- User management [`#189`](https://github.com/fireblocks/fireblocks-sdk-js/pull/189)
+
+#### [v5.13.0](https://github.com/fireblocks/fireblocks-sdk-js/compare/v5.12.0...v5.13.0)
+
+> 7 December 2023
+
+- Bump axios from 0.27.2 to 1.6.0 [`#231`](https://github.com/fireblocks/fireblocks-sdk-js/pull/231)
+- feat: ncw latest backup [`#225`](https://github.com/fireblocks/fireblocks-sdk-js/pull/225)
+- Bump semver from 5.7.1 to 5.7.2 [`#182`](https://github.com/fireblocks/fireblocks-sdk-js/pull/182)
+
+#### [v5.12.0](https://github.com/fireblocks/fireblocks-sdk-js/compare/v5.11.0...v5.12.0)
+
+> 28 November 2023
+
+- Added sourceWalletId for NCW under TransactionFilter [`#228`](https://github.com/fireblocks/fireblocks-sdk-js/pull/228)
+- Add get asset wallets paginated endpoint [`#150`](https://github.com/fireblocks/fireblocks-sdk-js/pull/150)
+
+#### [v5.11.0](https://github.com/fireblocks/fireblocks-sdk-js/compare/v5.10.0...v5.11.0)
 
 > 16 November 2023
 
 - staking-sdk: update staking sdk [`#222`](https://github.com/fireblocks/fireblocks-sdk-js/pull/222)
+- Revert mistake in updateVaultAccount [`#224`](https://github.com/fireblocks/fireblocks-sdk-js/pull/224)
+- Add getPaginatedAddresses method [`#223`](https://github.com/fireblocks/fireblocks-sdk-js/pull/223)
+
+#### [v5.10.0](https://github.com/fireblocks/fireblocks-sdk-js/compare/v5.9.0...v5.10.0)
+
+> 15 November 2023
+
+- contract-service - adding endpoints  [`#221`](https://github.com/fireblocks/fireblocks-sdk-js/pull/221)
+- added bulk functions [`#209`](https://github.com/fireblocks/fireblocks-sdk-js/pull/209)
+- Staking sdk [`#216`](https://github.com/fireblocks/fireblocks-sdk-js/pull/216)
 
 #### [v5.9.0](https://github.com/fireblocks/fireblocks-sdk-js/compare/v5.8.0...v5.9.0)